Changed BackEndHelperFunctions (and all methods in it) to be static. Changed the structure of one test
