+++ Tests/BackEndHelperFunctionsTests.cs
@@ -7,7 +7,6 @@ namespace YoCode_XUnit
 {
     public class BackEndHelperFunctionsTests
     {
-        internal BackEndHelperFunctions test;
 
         public List<double> expectedConversionResults;
         public List<double> expectedConversionInputs;
@@ -16,11 +15,9 @@ namespace YoCode_XUnit
         private readonly List<string> expectedActionLines;
         private readonly List<string> expectedActionLinesResult;
 
-        internal BackEndHelperFunctions Test { get => test; set => test = value; }
 
         public BackEndHelperFunctionsTests()
         {
-            Test = new BackEndHelperFunctions();
 
             expectedActionLines = new List<string>
             {
@@ -39,13 +36,15 @@ namespace YoCode_XUnit
             expectedConversionInputs = new List<double> { 1, 2, 5, 10, 20, 50 };
             mult = 2.54;
 
-            expectedConversionResults.Should().BeEquivalentTo(Test.MakeConversion(expectedConversionInputs, mult));
+            expectedConversionResults.Should().BeEquivalentTo(BackEndHelperFunctions.MakeConversion(expectedConversionInputs, mult));
         }
 
         [Fact]
         public void Test_ExtractActionsFromList()
         {
-            expectedActionLinesResult.Should().BeEquivalentTo(Test.ExtractActionsFromList(expectedActionLines,"value=\"","\""));
+            expectedActionLinesResult.Should().BeEquivalentTo(BackEndHelperFunctions.ExtractActionsFromList(expectedActionLines,"value=\"","\""));
+            BackEndHelperFunctions.ExtractActionsFromList(expectedActionLines,"value=\"","\"").Should().BeEquivalentTo(expectedActionLinesResult);
+
         }
 
         [Fact]
@@ -60,52 +59,6 @@ namespace YoCode_XUnit
             "Miles to kilometers".ToLower().ContainsAny(testMiToKmKeys).Should().BeTrue("testMiToKmKeys failed");
         }
 
-        [Fact]
-        public void Test_FindActualResultForExpectation()
-        {
-            var listOfActualResults = new List<UnitConverterResults>();
 
-            var actualElem1 = new UnitConverterResults
-            {
-                input = 2,
-                action = "Inches to Centimeters",
-                output = 5.08
-            };
-
-            var actualElem2 = new UnitConverterResults
-            {
-                input = 5,
-                action = "Miles to kilometers",
-                output = 8.04672
-            };
-
-            var actualElem3 = new UnitConverterResults
-            {
-                input = 20,
-                action = "Yards to meters",
-                output = 18.288
-            };
-
-            var testElem1 = new UnitConverterResults
-            {
-                input = 20,
-                action = "Yards to meters",
-                output = 6
-            };
-
-            var testElem2 = new UnitConverterResults
-            {
-                input = 5,
-                action = "Miles to kilometers",
-                output = 8.2
-            };
-
-            listOfActualResults.Add(actualElem1);
-            listOfActualResults.Add(actualElem2);
-            listOfActualResults.Add(actualElem3);
-
-            actualElem3.Should().BeEquivalentTo(UnitConverterCheck.FindActualResultForExpectation(actualElem3, listOfActualResults));
-            actualElem3.Should().BeEquivalentTo(UnitConverterCheck.FindActualResultForExpectation(testElem1, listOfActualResults));
-        }
     }
 }
